Update server.js
Accessing the date parameter from the route /api/timestamp/:date? correctly via req.params.date and not the date query string using req.query.date as I wrongly used before. I was confused about how to access a named route parameter, in this case the "date" parameter. The confusion stemmed from the date_string description in the user stories section.
I've cleaned up the code a bit and added a few comments for future reference and for a clearer picture.
The project is now complete, fully functional and ready to be submitted to FCC for review.